use session method to save current user

+++ controllers/notes/delete.php
@@ -3,8 +3,9 @@
 use App\Container;
 use App\Database ;
 use App\Validator; 
+use App\Session;
 
-$current_user = 1 ;
+$current_user =Session::get("auth_user_id") ?? 0; 
 
 $db = Container::resolve(Database::class);
 
@@ -12,12 +13,16 @@ $note_id = $_POST['id'];
 
 if($note_id >0){
     // check if note exists 
-    $note = $db->query("SELECT * FROM notes WHERE id = :id" , ["id" => $note_id])->findOrFail();
-  
-    authorize($note["user_id"] === $current_user) ;
-    // if authorized 
-    $db->query("DELETE FROM notes WHERE id = :id" , ["id" =>$note_id]);
-    
-    header("location: /notes");
+    if(Validator::exists($note_id , 'notes' , 'id')){
+        authorize($note["user_id"] === $current_user) ;
+        // if authorized 
+        $db->query("DELETE FROM notes WHERE id = :id" , ["id" =>$note_id]);
+        
+        header("location: /notes");
+       
+    }    
 }
+
+
+
